Coalesce some tests.
