--- libs/pavex_cli/tests/ui_tests/runtime_singletons_must_be_clonable/test_config.toml
@@ -1,6 +1,0 @@
-description = "Singletons must implement Clone"
-
-[expectations]
-codegen = "fail"
-
-
