--- libs/pavex_cli/tests/ui_tests/runtime_singletons_must_be_sync/test_config.toml
@@ -1,6 +1,0 @@
-description = "Singletons need to implement Sync"
-
-[expectations]
-codegen = "fail"
-
-
