+++ libs/pavex_cli/tests/ui_tests/trait_constraints_on_runtime_singletons/lib.rs
@@ -2,6 +2,28 @@ use std::rc::Rc;
 
 use pavex_builder::{f, AppBlueprint, Lifecycle};
 
+pub struct NonSendSingleton(Rc<()>);
+
+impl Clone for NonSendSingleton {
+    fn clone(&self) -> NonSendSingleton {
+        Self(Rc::clone(&self.0))
+    }
+}
+
+impl NonSendSingleton {
+    pub fn new() -> NonSendSingleton {
+        todo!()
+    }
+}
+
+pub struct NonCloneSingleton;
+
+impl NonCloneSingleton {
+    pub fn new() -> NonCloneSingleton {
+        todo!()
+    }
+}
+
 pub struct NonSyncSingleton(std::sync::mpsc::Sender<()>);
 
 impl Clone for NonSyncSingleton {
@@ -16,12 +38,18 @@ impl NonSyncSingleton {
     }
 }
 
-pub fn handler(_s: NonSyncSingleton) -> pavex_runtime::response::Response {
+pub fn handler(
+    _s: NonSendSingleton,
+    _a: NonSyncSingleton,
+    _c: NonCloneSingleton,
+) -> pavex_runtime::response::Response {
     todo!()
 }
 
 pub fn blueprint() -> AppBlueprint {
     let mut bp = AppBlueprint::new();
+    bp.constructor(f!(crate::NonSendSingleton::new), Lifecycle::Singleton);
+    bp.constructor(f!(crate::NonCloneSingleton::new), Lifecycle::Singleton);
     bp.constructor(f!(crate::NonSyncSingleton::new), Lifecycle::Singleton);
     // The handler is needed because bounds are only checked for singletons
     // that are used at runtime
