+++ libs/pavex_cli/tests/ui_tests/trait_constraints_on_runtime_singletons/test_config.toml
